# openapi.json
{
    "openapi": "3.1.0",
    "info": {
        "title": "Class Chatting API",
        "version": "0.0.1",
        "description": "Dokumen OpenAPI 3.1 untuk Class Chatting.\n\nDokumentasi ini mencakup:\n- Public API di bawah prefix `/api`\n- Private API di bawah prefix `/private-api/{api_key}`\n- Firebase webhook di bawah prefix `/api/firebase`\n\nGunakan `/docs/api` untuk dokumentasi interaktif, `/docs/api.json` untuk OpenAPI JSON runtime, dan `openapi.json` untuk artefak yang dapat dikonsumsi AI, generator SDK, atau automation tooling."
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Current application"
        }
    ],
...
    "paths": {
...
        "/private-api/{api_key}/video/list": {
            "get": {
                "operationId": "privateVideosList",
                "description": "Mengambil daftar video private. Request memakai `page` berbasis 1, tetapi field `page` pada response saat ini mengikuti implementasi legacy dan bernilai indeks berbasis 0.",
                "summary": "List private videos",
                "tags": [
                    "Private Videos"
                ],
                "parameters": [
                    {
                        "name": "api_key",
                        "in": "path",
                        "required": true,
                        "description": "Private API key yang wajib dikirim pada path URL.",
                        "schema": {
                            "type": "string"
                        },
                        "example": "test-private-key"
                    },
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Nomor halaman request. Nilai minimum efektif adalah 1.",
                        "schema": {
                            "type": "integer"
                        },
                        "example": 1
                    },
                    {
                        "name": "limit",
                        "in": "query",
                        "description": "Jumlah video per halaman.",
                        "schema": {
                            "type": "integer"
                        },
                        "example": 20
                    },
                    {
                        "name": "search",
                        "in": "query",
                        "description": "Filter judul video berbasis kata kunci.",
                        "schema": {
                            "type": "string"
                        },
                        "example": "tajwid"
                    },
                    {
                        "name": "video",
                        "in": "query",
                        "description": "Flag legacy untuk filtering tambahan pada implementasi lama.",
                        "schema": {
                            "type": "string"
                        },
                        "example": "1"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "page": {
                                            "type": [
                                                "object",
                                                "null"
                                            ]
                                        },
                                        "limit": {
                                            "type": "integer"
                                        },
                                        "search": {
                                            "anyOf": [
                                                {
                                                    "type": "string"
                                                },
                                                {
                                                    "type": "array",
                                                    "items": {}
                                                },
                                                {
                                                    "type": "null"
                                                }
                                            ]
                                        },
                                        "list_video": {
                                            "type": "string"
                                        },
                                        "slug": {
                                            "type": "string",
                                            "const": "list"
                                        },
                                        "status": {
                                            "type": "string",
                                            "const": "success"
                                        }
                                    },
                                    "required": [
                                        "page",
                                        "limit",
                                        "search",
                                        "list_video",
                                        "slug",
                                        "status"
                                    ]
                                }
                            }
                        }
                    }
                }
            }
        },
        "/private-api/{video_id}/video/detail/{video_id}": {
            "get": {
                "operationId": "privateVideosShow",
                "description": "Mengambil detail satu video private berdasarkan `video_id` legacy.",
                "summary": "Get private video detail",
                "tags": [
                    "Private Videos"
                ],
                "parameters": [
                    {
                        "name": "api_key",
                        "in": "path",
                        "required": true,
                        "description": "Private API key yang wajib dikirim pada path URL.",
                        "schema": {},
                        "example": "test-private-key"
                    },
                    {
                        "name": "video_id",
                        "in": "path",
                        "required": true,
                        "description": "Identifier legacy video yang dipakai endpoint private.",
                        "schema": {
                            "type": "string"
                        },
                        "example": "abc123def45"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "string"
                                        },
                                        "slug": {
                                            "type": "string",
                                            "const": "detail"
                                        },
                                        "status": {
                                            "type": "string",
                                            "const": "success"
                                        }
                                    },
                                    "required": [
                                        "data",
                                        "slug",
                                        "status"
                                    ]
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "const": "error"
                                        },
                                        "message": {
                                            "type": "string",
                                            "const": "Video not Found"
                                        }
                                    },
                                    "required": [
                                        "status",
                                        "message"
                                    ]
                                }
                            }
                        }
                    }
                }
            }
        },
        "/private-api/{api_key}/video/add": {
            "post": {
                "operationId": "privateVideosCreate",
                "description": "Membuat data video private baru. Implementasi saat ini menerima body JSON secara fleksibel dan akan mengembalikan payload yang sama beserta `video_id` dan `date_upload` yang dibuat server.",
                "summary": "Create private video",
                "tags": [
                    "Private Videos"
                ],
                "parameters": [
                    {
                        "name": "api_key",
                        "in": "path",
                        "required": true,
                        "description": "Private API key yang wajib dikirim pada path URL.",
                        "schema": {
                            "type": "string"
                        },
                        "example": "test-private-key"
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object",
                                "properties": {
                                    "title": {},
                                    "slug": {},
                                    "description": {},
                                    "uploaded_by": {
                                        "type": "integer",
                                        "description": "ID user pengunggah.",
                                        "example": 1
                                    },
                                    "thumbnail": {},
                                    "video_url": {},
                                    "storage_path": {},
                                    "file_path": {},
                                    "provider": {},
                                    "metadata": {
                                        "type": "object",
                                        "description": "Metadata tambahan dalam bentuk objek JSON.",
                                        "example": {
                                            "duration": 120,
                                            "resolution": "1280x720"
                                        },
                                        "additionalProperties": {}
                                    },
                                    "tags": {
                                        "type": "array",
                                        "description": "Daftar tag video.",
                                        "example": [
                                            "kelas-4",
                                            "tajwid"
                                        ],
                                        "items": {
                                            "type": "string"
                                        }
                                    }
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "slug": {
                                            "type": "string",
                                            "const": "add"
                                        },
                                        "data": {
                                            "type": "string"
                                        },
                                        "status": {
                                            "type": "string",
                                            "const": "success"
                                        }
                                    },
                                    "required": [
                                        "slug",
                                        "data",
                                        "status"
                                    ]
                                }
                            }
                        }
                    }
                }
            }
        },
        "/private-api/{video_id}/video/update/{video_id}": {
            "put": {
                "operationId": "privateVideosUpdate",
                "description": "Memperbarui data video private berdasarkan `video_id` legacy. Body JSON yang dikirim akan dipakai langsung untuk operasi update.",
                "summary": "Update private video",
                "tags": [
                    "Private Videos"
                ],
                "parameters": [
                    {
                        "name": "api_key",
                        "in": "path",
                        "required": true,
                        "description": "Private API key yang wajib dikirim pada path URL.",
                        "schema": {},
                        "example": "test-private-key"
                    },
                    {
                        "name": "video_id",
                        "in": "path",
                        "required": true,
                        "description": "Identifier legacy video yang dipakai endpoint private.",
                        "schema": {
                            "type": "string"
                        },
                        "example": "abc123def45"
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object",
                                "properties": {
                                    "title": {},
                                    "slug": {},
                                    "description": {},
                                    "uploaded_by": {
                                        "type": "integer",
                                        "description": "ID user pengunggah.",
                                        "example": 1
                                    },
                                    "thumbnail": {},
                                    "video_url": {},
                                    "storage_path": {},
                                    "file_path": {},
                                    "provider": {},
                                    "metadata": {
                                        "type": "object",
                                        "description": "Metadata tambahan dalam bentuk objek JSON.",
                                        "example": {
                                            "duration": 120,
                                            "resolution": "1280x720"
                                        },
                                        "additionalProperties": {}
                                    },
                                    "tags": {
                                        "type": "array",
                                        "description": "Daftar tag video.",
                                        "example": [
                                            "kelas-4",
                                            "tajwid"
                                        ],
                                        "items": {
                                            "type": "string"
                                        }
                                    }
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "slug": {
                                            "type": "string",
                                            "const": "update"
                                        },
                                        "data": {
                                            "type": "string"
                                        },
                                        "status": {
                                            "type": "string",
                                            "const": "success"
                                        }
                                    },
                                    "required": [
                                        "slug",
                                        "data",
                                        "status"
                                    ]
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "const": "error"
                                        },
                                        "message": {
                                            "type": "string",
                                            "const": "Update Failed"
                                        }
                                    },
                                    "required": [
                                        "status",
                                        "message"
                                    ]
                                }
                            }
                        }
                    }
                }
            }
        },
        "/private-api/{video_id}/video/delete/{video_id}": {
            "delete": {
                "operationId": "privateVideosDelete",
                "description": "Menghapus video private berdasarkan `video_id` legacy.",
                "summary": "Delete private video",
                "tags": [
                    "Private Videos"
                ],
                "parameters": [
                    {
                        "name": "api_key",
                        "in": "path",
                        "required": true,
                        "description": "Private API key yang wajib dikirim pada path URL.",
                        "schema": {},
                        "example": "test-private-key"
                    },
                    {
                        "name": "video_id",
                        "in": "path",
                        "required": true,
                        "description": "Identifier legacy video yang dipakai endpoint private.",
                        "schema": {
                            "type": "string"
                        },
                        "example": "abc123def45"
                    }
                ],
                "responses": {
                    "500": {
                        "description": "",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "const": "error"
                                        },
                                        "message": {
                                            "type": "string",
                                            "const": "Delete Failed"
                                        }
                                    },
                                    "required": [
                                        "status",
                                        "message"
                                    ]
                                }
                            }
                        }
                    },
                    "200": {
                        "description": "",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "slug": {
                                            "type": "string",
                                            "const": "add"
                                        },
                                        "msg": {
                                            "type": "string",
                                            "const": "Delete video succesfully"
                                        },
                                        "status": {
                                            "type": "string",
                                            "const": "success"
                                        }
                                    },
                                    "required": [
                                        "slug",
                                        "msg",
                                        "status"
                                    ]
                                }
                            }
                        }
                    }
                }
            }
        }
    },
...
}